New file: lib/utils/encoding.ts — single home for the slot serialization format:   - encodeSlots — converts a slot structure to the comma-separated action ID string (moved from slots.ts)   - decodeSlots — parses that string back to an array (extracted from the inline split(','))   - ENCODED_SLOTS_PARAM = 's' — the canonical URL param key   - ENCODED_SLOTS_PARAM_ALIASES — ['s', 's1', 'encodedSlots'] — the historical aliases that need normalizing   - readSlotParam — resolves s1 || s to a single value (extracted from inline in url.ts)
  lib/utils/slots.ts — removed encodeSlots and assignActionIds (implementation details of encoding), imports
  them from encoding.ts. Uses decodeSlots() instead of bare .split(',').

  lib/utils/url.ts — replaced the inline string arrays and query.s1 || query.s with the named constants and
  readSlotParam. Now if the URL format ever changes (new alias, different separator, base64, etc.) there's one
  file to edit.
